requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
attrs==24.2.0; python_version >= '3.7'
backports.zoneinfo==0.2.1; python_version < '3.9'
django==4.2; python_version >= '3.8'
django-configurations==2.5.1; python_version >= '3.8' and python_version < '4.0'
django-extensions==3.2.3; python_version >= '3.6'
django-rest-knox==5.0.1; python_version >= '3.6'
djangorestframework==3.15.2; python_version >= '3.8'
drf-spectacular==0.27.2; python_version >= '3.7'
importlib-resources==6.4.5; python_version < '3.9'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
pkgutil-resolve-name==1.3.10; python_version < '3.9'
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.35.1; python_version >= '3.8'
rpds-py==0.20.0; python_version >= '3.8'
sqlparse==0.5.1; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.10'
uritemplate==4.1.1; python_version >= '3.6'
zipp==3.20.2; python_version < '3.10'
